requirements.txt
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
scipy==1.11.1
matplotlib==3.7.2
tqdm==4.65.0
scikit-learn==1.3.0
ccxt==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
python-dotenv==1.0.1
bcrypt==4.0.1
yagmail==0.15.293